requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dm-tree==0.1.8
docstring_parser==0.16
etils==1.10.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
imageio==2.36.0
immutabledict==4.2.0
importlib_resources==6.4.5
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.3
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.13.2
networkx==3.4.2
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
promise==2.3
protobuf==4.25.5
psutil==6.1.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
simple-parsing==0.1.6
six==1.16.0
smmap==5.0.1
streamlit==1.40.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.7
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
termcolor==2.5.0
tifffile==2024.9.20
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.1.1
wrapt==1.16.0
zipp==3.20.2
